Asylum Applications

Asylum claims submitted by year and countries of asylum and origin. Claims are submitted by asylum applicants and are applications for international protection.

Usage

asylum_applications

Format

A tibble.

year

The year (e.g. 2020).

coo_name

Country of origin name.

coo

Country of origin UNHCR code.

coo_iso

Country of origin ISO code.

coa_name

Country of asylum name.

coa

Country of asylum UNHCR code.

coa_iso

Country of asylum ISO code.

procedure_type

The procedure type describes the authority with whom the asylum claim was lodged. It can be one of: G – Government, J - Joint, U – UNHCR.

app_type

The application type describes the stage of the procedure. Common values are N (New), R (Repeat) and A (Appeal).

dec_level

The decision level provides more detail of the stage of the procedure. Common values are NA (New applications), RA (repeat/reopened applications) and AR (administrative review).

app_pc

The data type describes whether the data provided reflects cases or persons. The most common situation is for the data on asylum applications lodged to reflect persons, but in certain situations (e.g USA), the data provided by governments is at the level of cases. A case can contain information on one or more persons.

applied

The total number of asylum claims for the given dataset.

Source

UNHCR Refugee Data Finder.

Details

See the notes on the data structure in the methodology pages for more details.

Examples

asylum_applications
#> # A tibble: 114,588 × 12
#>     year coo_name   coo   coo_iso coa_name coa   coa_iso procedure_type app_type
#>    <dbl> <chr>      <chr> <chr>   <chr>    <chr> <chr>   <chr>          <chr>   
#>  1  2006 Afghanist… AFG   AFG     Austral… AUL   AUS     G              A       
#>  2  2006 Albania    ALB   ALB     Austral… AUL   AUS     G              A       
#>  3  2006 Algeria    ALG   DZA     Austral… AUL   AUS     G              A       
#>  4  2006 Egypt      ARE   EGY     Austral… AUL   AUS     G              A       
#>  5  2006 Bahrain    BAH   BHR     Austral… AUL   AUS     G              A       
#>  6  2006 Bangladesh BGD   BGD     Austral… AUL   AUS     G              A       
#>  7  2006 Belarus    BLR   BLR     Austral… AUL   AUS     G              A       
#>  8  2006 Brazil     BRA   BRA     Austral… AUL   AUS     G              A       
#>  9  2006 Bulgaria   BUL   BGR     Austral… AUL   AUS     G              A       
#> 10  2006 Cambodia   CAM   KHM     Austral… AUL   AUS     G              A       
#> # ℹ 114,578 more rows
#> # ℹ 3 more variables: dec_level <chr>, app_pc <chr>, applied <dbl>