IDMC Data

Data on Internally displaced persons due to conflict and violence that is produced by the Internal Displacement Monitoring Centre.

Usage

idmc

Format

A tibble.

year

The year (e.g. 2020).

coo_name

Country of origin name.

coo

Country of origin UNHCR code.

coo_iso

Country of origin ISO code.

coa_name

Country of asylum name.

coa

Country of asylum UNHCR code.

coa_iso

Country of asylum ISO code.

total

The total number of internally displaced persons due to conflict and violence.

Source

UNHCR Refugee Data Finder.

Details

See the notes on the data structure in the methodology pages for more details.

Examples

idmc
#> # A tibble: 881 × 8
#>     year coo_name coo   coo_iso coa_name coa   coa_iso    total
#>    <dbl> <chr>    <chr> <chr>   <chr>    <chr> <chr>      <dbl>
#>  1  1990 Unknown  UKN   UNK     Unknown  UKN   UNK     21300000
#>  2  1991 Unknown  UKN   UNK     Unknown  UKN   UNK     22500000
#>  3  1992 Unknown  UKN   UNK     Unknown  UKN   UNK     25000000
#>  4  1993 Unknown  UKN   UNK     Unknown  UKN   UNK     27500000
#>  5  1994 Unknown  UKN   UNK     Unknown  UKN   UNK     28000000
#>  6  1995 Unknown  UKN   UNK     Unknown  UKN   UNK     22000000
#>  7  1996 Unknown  UKN   UNK     Unknown  UKN   UNK     19700000
#>  8  1997 Unknown  UKN   UNK     Unknown  UKN   UNK     17400000
#>  9  1998 Unknown  UKN   UNK     Unknown  UKN   UNK     19300000
#> 10  1999 Unknown  UKN   UNK     Unknown  UKN   UNK     21300000
#> # ℹ 871 more rows