Population figures

UNHCR data on displacement at the end of the year.

Usage

population

Format

A tibble.

year

The year (e.g. 2020).

coo_name

Country of origin name.

coo

Country of origin UNHCR code.

coo_iso

Country of origin ISO code.

coa_name

Country of asylum name.

coa

Country of asylum UNHCR code.

coa_iso

Country of asylum ISO code.

refugees

The number of refugees.

asylum_seekers

The number of asylum-seekers.

returned_refugees

The number of returned refugees.

idps

The number of internally displaced persons.

returned_idps

The number of returned internally displaced persons.

stateless

The number of stateless persons.

ooc

The number of others of concern to UNHCR.

oip

The number of other people in need of international protection.

hst

The number of host community members.

Source

UNHCR Refugee Data Finder.

Details

See the notes on the data structure in the methodology pages for more details.

Examples

population
#> # A tibble: 132,491 × 16
#>     year coo_name coo   coo_iso coa_name   coa   coa_iso refugees asylum_seekers
#>    <dbl> <chr>    <chr> <chr>   <chr>      <chr> <chr>      <dbl>          <dbl>
#>  1  1951 Unknown  UKN   UNK     Australia  AUL   AUS       180000              0
#>  2  1951 Unknown  UKN   UNK     Austria    AUS   AUT       282000              0
#>  3  1951 Unknown  UKN   UNK     Belgium    BEL   BEL        55000              0
#>  4  1951 Unknown  UKN   UNK     Canada     CAN   CAN       168511              0
#>  5  1951 Unknown  UKN   UNK     Denmark    DEN   DNK         2000              0
#>  6  1951 Unknown  UKN   UNK     France     FRA   FRA       290000              0
#>  7  1951 Unknown  UKN   UNK     United Ki… GBR   GBR       208000              0
#>  8  1951 Unknown  UKN   UNK     Germany    GFR   DEU       265000              0
#>  9  1951 Unknown  UKN   UNK     Greece     GRE   GRC        18000              0
#> 10  1951 Unknown  UKN   UNK     China, Ho… HKG   HKG        30000              0
#> # ℹ 132,481 more rows
#> # ℹ 7 more variables: returned_refugees <dbl>, idps <dbl>, returned_idps <dbl>,
#> #   stateless <dbl>, ooc <dbl>, oip <dbl>, hst <dbl>