Asylum Decisions

Decisions taken on asylum claims by year and countries of asylum and origin. Asylum claims are applications for international protection and decisions on asylum claims can be positive, negative or otherwise closed.

Usage

asylum_decisions

Format

A tibble.

year

The year (e.g. 2020).

coo_name

Country of origin name.

coo

Country of origin UNHCR code.

coo_iso

Country of origin ISO code.

coa_name

Country of asylum name.

coa

Country of asylum UNHCR code.

coa_iso

Country of asylum ISO code.

procedure_type

The procedure type describes the authority with whom the asylum claim was lodged. It can be one of: G – Government, J - Joint, U – UNHCR.

dec_level

The decision level provides more detail of the stage of the procedure. Common values are NA (New applications), RA (repeat/reopened applications) and AR (administrative review).

dec_pc

The data type describes whether the data provided reflects cases or persons. The most common situation is for the data on asylum applications lodged to reflect persons, but in certain situations (e.g USA), the data provided by governments is at the level of cases. A case can contain information on one or more persons.

dec_recognized

The number of asylum applicants recognised as refugees under UNHCR’s mandate.

dec_other

The number of asylum applicants recognised as refugees under other forms of complementary protection.

dec_rejected

The number of asylum applicants whose asylum claims have been rejected.

dec_closed

The number of asylum applicants whose asylum claims have been otherwise closed.

dec_total

The total number of asylum applicants for which a decision on their asylum claim has been made in the given reference period.

Source

UNHCR Refugee Data Finder.

Details

See the notes on the data structure in the methodology pages for more details.

Examples

asylum_decisions
#> # A tibble: 107,396 × 15
#>     year coo_name  coo   coo_iso coa_name coa   coa_iso procedure_type dec_level
#>    <dbl> <chr>     <chr> <chr>   <chr>    <chr> <chr>   <chr>          <chr>    
#>  1  2000 Afghanis… AFG   AFG     Austral… AUL   AUS     G              AR       
#>  2  2000 Albania   ALB   ALB     Austral… AUL   AUS     G              AR       
#>  3  2000 Algeria   ALG   DZA     Austral… AUL   AUS     G              AR       
#>  4  2000 Egypt     ARE   EGY     Austral… AUL   AUS     G              AR       
#>  5  2000 Armenia   ARM   ARM     Austral… AUL   AUS     G              AR       
#>  6  2000 Banglade… BGD   BGD     Austral… AUL   AUS     G              AR       
#>  7  2000 Brazil    BRA   BRA     Austral… AUL   AUS     G              AR       
#>  8  2000 Bosnia a… BSN   BIH     Austral… AUL   AUS     G              AR       
#>  9  2000 Bulgaria  BUL   BGR     Austral… AUL   AUS     G              AR       
#> 10  2000 Cambodia  CAM   KHM     Austral… AUL   AUS     G              AR       
#> # ℹ 107,386 more rows
#> # ℹ 6 more variables: dec_pc <chr>, dec_recognized <dbl>, dec_other <dbl>,
#> #   dec_rejected <dbl>, dec_closed <dbl>, dec_total <dbl>