Demographics

Demographic and sub-national data, where such disaggregation is available. UNHCR collects this information for all population types, as well as two durable solutions (returned IDPs and refugees). Demographics are also available for IDMC and UNRWA datasets.

Usage

demographics

Format

A tibble.

year

The year (e.g. 2020).

coo_name

Country of origin name.

coo

Country of origin UNHCR code.

coo_iso

Country of origin ISO code.

coa_name

Country of asylum name.

coa

Country of asylum UNHCR code.

coa_iso

Country of asylum ISO code.

pop_type

The three character code specifying the population type (e.g. REF for refugees, ASY for asylum-seekers etc).

location

The textual description of the location.

loc_type

An indication of whether the location is urban or rural. An urban location is classified as a settlement with more than 5,000 inhabitants. It can be one of the following values: C – Camp. Note that C was discontinued after 2010, at which point the accommodation type became available, U – Urban, R – Rural, V - Various/unknown

acc_type

Accommodation Type became available in 2010 but is only populated from 2012. The code represents one of: I - Individual accommodation, S - Self-settled camp, P - Planned/managed camp, C - Collective centre, R - Reception/transit camp, U - Undefined

f_0_4

The number of female children aged 0 to 4 inclusive.

f_5_11

The number of female children aged 5 to 11 inclusive.

f_12_17

The number of female children aged 12 to 17 inclusive.

f_18_59

The number of female adults aged 18 to 59 inclusive.

f_60

The number of female adults aged 60 or more.

f_other

The number of females for whom age disaggregation is unavailable.

f_total

The total number of females (where disaggregation by sex is available).

m_0_4

The number of male children aged 0 to 4 inclusive.

m_5_11

The number of male children aged 5 to 11 inclusive.

m_12_17

The number of male children aged 12 to 17 inclusive.

m_18_59

The number of male adults aged 18 to 59 inclusive.

m_60

The number of male adults aged 60 or more.

m_other

The number of males for whom age disaggregation is unavailable.

m_total

The total number of males (where disaggregation by sex is available).

total

The total number of persons.

Source

UNHCR Refugee Data Finder.

Details

See the notes on the data structure in the methodology pages for more details.

Examples

demographics
#> # A tibble: 209,731 × 26
#>     year coo_name         coo   coo_iso coa_name coa   coa_iso pop_type location
#>    <dbl> <chr>            <chr> <chr>   <chr>    <chr> <chr>   <chr>    <chr>   
#>  1  2001 Afghanistan      AFG   AFG     Afghani… AFG   AFG     IDP      Central 
#>  2  2001 Afghanistan      AFG   AFG     Afghani… AFG   AFG     IDP      Dispers…
#>  3  2001 Afghanistan      AFG   AFG     Afghani… AFG   AFG     IDP      North   
#>  4  2001 Afghanistan      AFG   AFG     Afghani… AFG   AFG     IDP      West    
#>  5  2001 Afghanistan      AFG   AFG     Afghani… AFG   AFG     RET      Dispers…
#>  6  2001 Iran (Islamic R… IRN   IRN     Afghani… AFG   AFG     REF      Kabul   
#>  7  2001 Iraq             IRQ   IRQ     Afghani… AFG   AFG     REF      Kabul   
#>  8  2001 Iraq             IRQ   IRQ     Albania  ALB   ALB     REF      Tirana  
#>  9  2001 North Macedonia  MCD   MKD     Albania  ALB   ALB     ASY      Tirana  
#> 10  2001 North Macedonia  MCD   MKD     Albania  ALB   ALB     REF      Tirana  
#> # ℹ 209,721 more rows
#> # ℹ 17 more variables: loc_type <chr>, acc_type <chr>, f_0_4 <dbl>,
#> #   f_5_11 <dbl>, f_12_17 <dbl>, f_18_59 <dbl>, f_60 <dbl>, f_other <dbl>,
#> #   f_total <dbl>, m_0_4 <dbl>, m_5_11 <dbl>, m_12_17 <dbl>, m_18_59 <dbl>,
#> #   m_60 <dbl>, m_other <dbl>, m_total <dbl>, total <dbl>